Added redirects and customized 404 page
Due to the new URL, the Algolia search is not working.
The URLs provided in the search results have the wrong "path" to the docs.

A case has been opened with Algolia and until it's fixed, as a workaround here the temporary setup:

- Added the `redirect` plugin to Docusaurus
- Added all the pages redirects in the `docusaurus.config.js`
- Modified the "page not found" (404) error message to display a button to refresh the page.

The 404 error message change is needed, as when a user clicks on a search result, Docusaurus doesn't trigger the redirect. Instead, the URL needs to be "retriggered" (page refresh) for the redirect to work.

Once the search results are updated with the new crawler configuration done by Algolia, this workaround will be retracted.

Signed-off-by: Nuno do Carmo <nuno.carmo@suse.com>