New quickstart UI and browsing
This PR changes the Quickstart menu and adds a new Quickstart UI page.

The menu shows the two options to install and use Elemental:

* First option is through Rancher Manager with the Elemental plugin
* Second option is with the CLI only

To avoid duplicates of common parts between the two UIs, the `partial` feature from Docusaurus is used.

This allows to write one portion of documentation into one page only, and import the page into both Quickstart pages.

Finally, the Overview has also been updated with the description of what's the Elemental stack. This is best fit there, instead of the Quickstart.

Signed-off-by: Nuno do Carmo <nuno.carmo@suse.com>