Versioning updated
The current versioning of only 2 branches, `Next` and `Stable`, is no more adequate.

This PR changes the versioning to a numeric one and will follow the `major`.`minor` versioning of [Elemental releases](https://github.com/rancher/elemental/releases).

The following changes have been applied in this specific order:

1. Renamed `Stable` to `1.0` in the `versions.js` and the `versioned_docs` directory

2. Created a new version from the existing `/docs` as the new version `1.1`

The version `1.0` will be worked on to reflect the features that only existed before the version 1.1.

Signed-off-by: Nuno do Carmo nuno.carmo@suse.com