Remove mentions of Teal
Elemental Teal has been the working name for the official elemental OS
built in OBS.

As more and more features are migrated into SLE Micro, Elemental Teal is
being switched out for SLE Micro directly.

This commit serves to make this more obvious in the docs and removes
alot of mentions of Elemental Teal.

Signed-off-by: Fredrik Lönnegren <fredrik.lonnegren@suse.com>